elt_doc_sad/pyproject.toml
[project]
name = "elt-doc-sad"
version = "0.1.0"
description = "Generate Solution Architecture Definition (SAD) documents from template"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-docx>=1.1.0",
    "openpyxl>=3.1.5",
]

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
